Add support for pre-Haswell architecture to clz64
The previous host-side code for clz64 required the ABM / BMI1
instruction set for lzclr; this was introduced in Intel Haswell (2013)
or AMD Piledriver (2012).  It would not issue an error on failure.
Since clz64 is in the code path for std::ostream &
operator<<(std::ostream & out, uint128_t x), this makes it possible to
print incorrect numbers.

The new version will use the gcc or clang builtin clzll, if they're
available, or else use the widely-available bsr instruction.