Improve last patch by giving the compiler more flexibility
The following only applies if the compiler doesn't have a builtin
clzll.  I don't know if that's even a supported configuration.

The previous version required the operand to be in memory; this
version lets it be in a register, which may be better for inlining.
Also, I neglected to include the clobber registers in the previous
version.